requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Markdown==3.5.2
PyYAML==6.0.1
python-frontmatter==1.0.1
Werkzeug==3.0.1
beautifulsoup4==4.12.2
firebase-admin==6.4.0 